feat(animated-beam-graph): add AnimatedBeamGraph component to visualize AI infrastructure connections fix(content): adjust grid layout and logo carousel column count for better responsiveness
style(features.tsx): refactor CardContent to use flexbox for better alignment and centering of content

refactor(hero.tsx): simplify HeroSection component by removing unused code and replacing it with AnimatedBeamGraph for better maintainability and readability
